+++ Src/libCZI_UnitTests/test_CziSubBlockDirectory.cpp
@@ -5,6 +5,7 @@
 #include "include_gtest.h"
 #include "testImage.h"
 #include "inc_libCZI.h"
+#include <cstddef>
 
 using namespace libCZI;
 
@@ -206,7 +207,7 @@ TEST(CziSubBlockDirectory, CziSubBlockDirectory1)
     };
 
     CCziSubBlockDirectory subBlkDir;
-    for (int i = 0; i < sizeof(data) / sizeof(data[0]); ++i)
+    for (size_t i = 0; i < sizeof(data) / sizeof(data[0]); ++i)
     {
         auto entry = SubBlkEntryFromSubBlockEntryData(data + i);
         subBlkDir.AddSubBlock(entry);
